+++ .gitignore
@@ -1,5 +1,6 @@
 #ignorar la carpeta .data para evitar quilombos
 #ademas de los archivos que no importa si se modifican.
+.data
 ServerMensajes/.data
 ServerMensajes/Client/model
 ServerMensajes/Client/deploy
