+++ ServerMensajes/Client/src/modelo_m/RearmadorLista.java
@@ -1,4 +1,4 @@
-package client;
+package modelo_m;
 
 import java.util.ArrayList;
