+++ ServerMensajes/Client/src/modelo_m/SistemaM.java
@@ -1,32 +1,34 @@
-package client;
+package modelo_m;
 
-import configuracion.Configuracion;
+import modelo_m.agenda.Agenda;
 
-import configuracion.LectorConfiguracion;
+import modelo_m.configuracion.Configuracion;
 
-import configuracion.NoLecturaConfiguracionException;
+import modelo_m.configuracion.LectorConfiguracion;
+
+import modelo_m.configuracion.NoLecturaConfiguracionException;
 
 import java.io.IOException;
 
-public class Sistema {
-    private static Sistema sistema;
+public class SistemaM {
+    private static SistemaM sistema;
     private static String ARCH_CONFIG = "configuracion.txt";
     
     private SocketServer server;
     private Agenda agenda;
     private Configuracion config;
     
-    private Sistema() throws NoLecturaConfiguracionException {
-        config = LectorConfiguracion.leerConfig(Sistema.ARCH_CONFIG);
+    private SistemaM() throws NoLecturaConfiguracionException {
+        config = LectorConfiguracion.leerConfig(SistemaM.ARCH_CONFIG);
         agenda = new Agenda();
         server = new SocketServer(agenda);
         iniciarServer();
     }
     
-    public static synchronized Sistema getInstance() {
+    public static synchronized SistemaM getInstance() {
         if (sistema == null) {
             try {
-                sistema = new Sistema();
+                sistema = new SistemaM();
             } catch (NoLecturaConfiguracionException e) {
                 System.out.println("Error al leer el archivo de configuracion. Por favor reinicie el programa.");
             }
