+++ ServerMensajes/Client/src/modelo_m/agenda/Usuario.java
@@ -1,4 +1,4 @@
-package client;
+package modelo_m.agenda;
 
 public class Usuario {
     private String nombre;
