+++ ServerMensajes/Client/src/modelo_m/configuracion/LectorConfiguracion.java
@@ -1,4 +1,4 @@
-package configuracion;
+package modelo_m.configuracion;
 
 import java.io.FileInputStream;
 import java.io.FileNotFoundException;
@@ -10,8 +10,8 @@ public class LectorConfiguracion {
      * @param nombreArch
      * Lee el archivo de configuracion.txt
      * y crea un objeto de clase <i>Configuracion</i> con su contenido.
-     * 
-     * @throws NoLecturaConfiguracionException
+     *
+     * @throws modelo_m.configuracion.NoLecturaConfiguracionException
      * Si ocurre un error con la lectura del archivo de configuracion.
      */
     public static Configuracion leerConfig(String nombreArch) throws NoLecturaConfiguracionException {
