+++ ServerMensajes/Client/src/modelo_m/configuracion/NoLecturaConfiguracionException.java
@@ -1,4 +1,4 @@
-package configuracion;
+package modelo_m.configuracion;
 
 public class NoLecturaConfiguracionException extends Exception {
     public NoLecturaConfiguracionException(String string, Throwable throwable, boolean b, boolean b1) {
