+++ ServerMensajes/Client/src/modelo_m/persistencia/IPersistencia.java
@@ -1,4 +1,4 @@
-package persistencia;
+package modelo_m.persistencia;
 
 import java.io.IOException;
