+++ ServerMensajes/Client/src/modelo_m/persistencia/PersistidorMensaje.java
@@ -1,4 +1,4 @@
-package persistencia;
+package modelo_m.persistencia;
 
 import java.io.BufferedReader;
 import java.io.BufferedWriter;
